+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All feature additions, significant bug fixes, and API changes will be documented
 in this file. This project follows [semantic versioning](https://semver.org/).
 
 ## 10.2
-- Add the `runCancelableThenPromise` to `Task`.
+- Add the `runCancelableThenPromise` method to `Task`.
+- Add the `onShutdown` method to `Cancelable`.
 - Fix an issue where the ordering and evaluation of guaranteed effects on an observable
   did not happen in a deterministic manner when using operators which transformed an
   `Observable` into a `Task` (`take`, `last`, `foreach`, and `foreachTask`). Now, any